multi-uv: fix warning

+++ docs/examples/multi-uv.c
@@ -177,7 +177,7 @@ static int cb_timeout(CURLM *multi, long timeout_ms,
     if(timeout_ms == 0)
       timeout_ms = 1; /* 0 means call curl_multi_socket_action asap but NOT
                          within the callback itself */
-    uv_timer_start(&uv->timeout, on_uv_timeout, timeout_ms,
+    uv_timer_start(&uv->timeout, on_uv_timeout, (uint64_t)timeout_ms,
                    0); /* do not repeat */
   }
   return 0;
